requirements.txt
Django==2.2.5
numpy==1.20.1
pandas==1.3.4
djangorestframework
markdown
django-filter
psycopg2-binary==2.8.6
dj_database_url
django-environ